analytics/mtmc_analytics/requirements.txt
kafka-python
pydantic
pymilvus
numpy
opencv-python
PyYAML